+++ setup.py
@@ -2,7 +2,7 @@
 import os
 from distutils.core import setup
 from setuptools import find_packages
-from src.version import __version__
+from genomic_address_service.version import __version__
 author = 'James Robertson'
 
 classifiers = """
@@ -24,25 +24,25 @@ def read(fname):
     return open(os.path.join(os.path.dirname(__file__), fname)).read()
 
 
-exec(open('cladeomatic/version.py').read())
+exec(open('genomic_address_service/version.py').read())
 
 setup(
-    name='profile_dists',
+    name='genomic_address_service',
     include_package_data=True,
     version=__version__,
     python_requires='>=3.8.2,<4',
     setup_requires=['pytest-runner'],
     tests_require=['pytest'],
     packages=find_packages(exclude=['tests']),
-    url='https://github.com/phac-nml/profile_dists',
+    url='https://github.com/phac-nml/genomic_address_service',
     license='GPLv3',
     author='James Robertson',
     author_email='james.robertson@phac-aspc.gc.ca',
     description=(
-        'Profile Dists: Rapid calcualtion of allele profile distances and distance base querying'),
-    keywords='cgMLST, wgMLST, outbreak, surveillance, clustering, distance matrix',
+        'Genomic Address Service: De novo clustering and cluster address assignment'),
+    keywords='cgMLST, wgMLST, outbreak, surveillance, clustering, nomenclature',
     classifiers=classifiers,
-    package_dir={'profile_dists': 'src'},
+    package_dir={'genomic_address_service': 'gas'},
     package_data={
         "": ["*.txt"],
     },
@@ -60,7 +60,7 @@ setup(
 
     entry_points={
         'console_scripts': [
-            'profile_dists=src.dist:main',
+            'genomic_address_service=genomic_address_service.main:main',
         ],
     },
 )
